Code factorization; File renaming
