+++ app/src/application/components/d3Blocks/conceptNav.less
@@ -10,8 +10,9 @@
 .conceptCircle {
     r: 8px;
     fill: @gray3;
-    stroke: @light-gray1;
+    stroke: @light-gray2;
     stroke-width: 0;
+    stroke-location: inside;
 }
 
 .toggle {
