+++ app/src/application/components/utils/conceptBar.tsx
